Realtime, Mothly Details, Comprehensive Reports, Product and Order APIs

+++ database/migrations/2023_10_29_181756_create_comprehensive_reports_table.php
@@ -16,11 +16,11 @@ return new class extends Migration
         Schema::create('comprehensive_reports', function (Blueprint $table) {
             $table->id();
             $table->string('remark');
-            $table->decimal('deposits');
-            $table->decimal('dispensing');
+            $table->integer('deposits');
+            $table->integer('dispensing');
             $table->integer('numffpeople');
             $table->integer('numapeople');
-            $table->decimal('profitloss');
+            $table->integer('profitloss');
             $table->timestamps();
         });
     }
